Working towards clients & servers communicating.
Updated TODO.md

Only pop commands from a player's command buffer if they existed in the
first place

Added doom_pprintf to print message to a specific player (kind of)

Added a clientside local command buffer, which is necessary because
there's a difference between the commands run locally and the commands
the server has acknowledged receipt of.

Wasn't initializing the network states module.

Wasn't sending the sync type in the setup message.

Player commands must include the playernum of the originating player,
because they are relayed (in command sync mode) by the server.

Copy commands from the network buffer directly to the command buffer
instead of using an intermediate, temporary command.

Fix a dumb copy/paste bug in n_proto.c in COMMAND_CLIENT_ONLY define.

Fix n_state; current_game_state is not a pointer and everything within
that module takes its reference.  Clearly written at 2am ;)

Players are fully serialized and deserialized now, instead of running a
memcpy.  This was necessary because commands needed to be serialized,
and they won't survive a memcpy.  Additionally memcpy'd data isn't
portable, so eventually all of these will have to be converted.