Getting close to complete sync
Gametic was inexplicably not included in savegames, adding it solved a
lot of problems (unsurprisingly)

Now resetting the blocklinks whenever a state is loaded

Fixed the server missing old states; was using `command_tic` to judge
instead of `state_tic`, so if a client's cmd was ahead of its state, the
server would remove states the client might need, and when it went back
to look for them, it had already deleted them.  So... just use
`state_tic` instead.